Add DietSessionSearch and FoodItemSearch Change some UI

+++ src/main/java/commands/diet/dietmanager/DietSessionClear.java
@@ -2,6 +2,7 @@ package commands.diet.dietmanager;
 
 import commands.Command;
 import storage.diet.Storage;
+import ui.diet.dietmanager.DietManagerUi;
 
 import java.io.File;
 import java.util.Objects;
@@ -10,7 +11,8 @@ import java.util.logging.Level;
 import static logger.SchwarzeneggerLogger.logger;
 
 public class DietSessionClear extends Command {
-    static final String FILEPATH = "saves/diet/";
+    private static final String FILEPATH = "saves/diet/";
+    private final DietManagerUi ui = new DietManagerUi();
 
     /**
      * Overrides execute for clear command to clear all diet sessions.
@@ -20,10 +22,14 @@ public class DietSessionClear extends Command {
     @Override
     public void execute(String input, Storage storage) {
         try {
-            File folder = new File(FILEPATH);
-            File[] listOfFiles = folder.listFiles();
-            for (int index = 0; index < Objects.requireNonNull(listOfFiles).length; index++) {
-                listOfFiles[index].delete();
+            if (checkConfirmation()) {
+                File folder = new File(FILEPATH);
+                File[] listOfFiles = folder.listFiles();
+                for (int index = 0; index < Objects.requireNonNull(listOfFiles).length; index++) {
+                    listOfFiles[index].delete();
+                }
+            } else {
+                System.out.println("You have aborted clear operation.");
             }
         } catch (NullPointerException e) {
             e.printStackTrace();
@@ -31,5 +37,12 @@ public class DietSessionClear extends Command {
         }
         System.out.println("You have cleared all diet sessions!");
         logger.log(Level.INFO, "Cleared all diet sessions");
+    }
+
+    private boolean checkConfirmation() {
+        System.out.println("\tAre you sure you want to clear all records? This action is irrevocable.");
+        System.out.println("\tKey in YES to confirm.");
+        String input = ui.getCommand("Diet Menu");
+        return  input.equals("YES");
     }
 }
